+++ examples/single_base_route/lib/main.dart
@@ -16,25 +16,32 @@ class MusicApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) => DeepLinkMaterialApp(
     // This is where the magic happens
-    navigation: (context) => Dispatcher()
+    navigation: Dispatcher()
       ..path<LibraryDL>(
-        (path) => LibraryPage(),
-        subNavigation: (context) => Dispatcher()
+        (route) => LibraryPage(),
+        subNavigation: Dispatcher()
           ..value<Artist, ArtistDL>(
-            (artist, path) => ArtistPage(artist: artist),
-            subNavigation: (context, artist) => Dispatcher()..value<Song, SongDL>((song, path) => SongPage(song: song)) // TODO: static extension method
+            (artist, route) => ArtistPage(artist: artist),
+            subNavigation: (artist) => Dispatcher()
+              ..song()
           )
           ..path<FavoritesDL>(
-            (path) => FavoritesPage(),
-            subNavigation: (context) => Dispatcher()..value<Song, SongDL>((song, path) => SongPage(song: song)) // TODO: static extension method
+            (route) => FavoritesPage(),
+            subNavigation: Dispatcher()
+              ..song()
           )
-          ..value<RouteNotFound, ErrorDL<RouteNotFound>>((exception, path) => ErrorPage(exception)),
+          ..value<RouteNotFound, ErrorDL<RouteNotFound>>((exception, route) => ErrorPage(exception)),
       )
       // Exception handling mappings and route dispatchers are specified independently
-      ..exception<RouteNotFound>((exception, path) => [LibraryDL(), ErrorDL<RouteNotFound>(exception)]),
+      ..exception<RouteNotFound>((exception, route) => [LibraryDL(), ErrorDL<RouteNotFound>(exception)]),
     defaultRoute: [LibraryDL()],
     splashScreen: SplashPage(),
     // Non-navigation related fields are still available
     themeMode: ThemeMode.light,
   );
+}
+
+/// Reusing code through static extension methods.
+extension DispatcherExtensions on Dispatcher {
+  void song() => value<Song, SongDL>((song, route) => SongPage(song: song));
 }